Core: Heap enhancements, NanoStrings, Fixed*Map, fix in Socket - non-ASCII char < 0 issue
git-svn-id: svn://ultimatepp.org/upp/trunk@6661 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Socket.cpp
@@ -746,7 +746,7 @@ int TcpSocket::Get_()
 	if(!IsOpen() || IsError() || IsEof() || IsAbort())
 		return -1;
 	ReadBuffer(GetEndTime());
-	return ptr < end ? *ptr++ : -1;
+	return ptr < end ? (byte)*ptr++ : -1;
 }
 
 int TcpSocket::Peek_(int end_time)
@@ -754,7 +754,7 @@ int TcpSocket::Peek_(int end_time)
 	if(!IsOpen() || IsError() || IsEof() || IsAbort())
 		return -1;
 	ReadBuffer(end_time);
-	return ptr < end ? *ptr : -1;
+	return ptr < end ? (byte)*ptr : -1;
 }
 
 int TcpSocket::Peek_()
